chore: rename directory to 'annotation' for clarity
